pyproject.toml
[project]
name = "loads"
version = "0.1.0"
description = "Car and freight tracking solution"
authors = [
    {name = "Alex Halitsky",email = "ceasione@gmail.com"}
]
license = {text = "MIT License"}
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "Flask (>=3.1.0,<4.0.0)",
    "flask-cors (>=5.0.0,<6.0.0)",
    "requests (>=2.0.0,<3.0.0)",
    "urllib3 (>=1.26.0,<1.27.0)",
    "python-telegram-bot (>=22.3,<23.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "psycopg[binary] (>=3.2.9,<4.0.0)",
    "testcontainers[postgresql] (>=4.12.0,<5.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pyngrok = "^7.0.0"
pytest-asyncio = "^1.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
